+++ tests/check_lastfmstats.py
@@ -1,0 +1,7 @@
+import memoryfm as mfm
+from pathlib import Path
+
+data_dir = Path(__file__).resolve().parent / "data"
+
+file = 
+