+++ tests/test_loaders.py
@@ -1,4 +1,0 @@
-#import pytest
-#from memoryfm.io._loaders import load_csv, load_json
-
-#def test_load_csv(